[🔥AUDIT🔥] Add job invoker role to job service account (#14)
🖍 _This is an audit!_ 🖍

## Summary:
This was missed in the initial implementation. If we want the service account ot be able to trigger a job, it needs permissions to do so!

Issue: INFRA-XXXX

## Test plan:
I'm testing this out on a culture-cron branch. Deploying and verifying that the scheduler works

Author: jwbron

Auditors: #infra-platform

Required Reviewers:

Approved By:

Checks: ✅ 1 check was successful

Pull Request URL: https://github.com/Khan/terraform-modules/pull/14