Store function archive bytes in the tfplan binary (#24)
## Summary:
I made a change already that I thought would store the archive bytes in the tfplan binary. However, after more testing it turns out that wasn't the case.

This change stores the archive file as a data field to ensure the bytes are calculated and included in the tfplan binary. This ensures the zip "file" is available when a plan is applied, which does not happen on the same machine as an apply.

Issue: INFRA-XXXX

## Test plan:
Run a plan with this branch on local in internal-webserver, delete the zip, and run an apply.

Author: jwbron

Reviewers: csilvers, jwbron

Required Reviewers:

Approved By: csilvers

Checks: ⏭️  2 checks have been skipped, ✅ 4 checks were successful

Pull Request URL: https://github.com/Khan/terraform-modules/pull/24