Update scheduled job and function to depend on secret access (#28)
## Summary:
Currently, the first `apply` after adding a new secret will often fail because the secret access is created as the same time as the job or function, so the job or function can fail to start due to lacking access to the secret. Having the job or function depend on the secred access should resolve that.

Issue: INFRA-10752

## Test plan:
I ran an apply with this on local in culture-cron and did no see any issues. Testing the actual timing issue is a bit tricker, so we'll just see if it comes up again.

Author: jwbron

Reviewers: copilot-pull-request-reviewer[bot], csilvers

Required Reviewers:

Approved By: csilvers

Checks: ✅ 4 checks were successful, ⏭️  1 check has been skipped

Pull Request URL: https://github.com/Khan/terraform-modules/pull/28